build fixes
svn:sync-xref-src-repo-rev: r85932

+++ src/VBox/Runtime/common/string/strcache.cpp
@@ -1,4 +1,4 @@
-/* $Id: strcache.cpp 46211 2013-05-22 11:00:02Z knut.osmundsen@oracle.com $ */
+/* $Id: strcache.cpp 46212 2013-05-22 11:04:30Z knut.osmundsen@oracle.com $ */
 /** @file
  * IPRT - String Cache.
  */
@@ -280,11 +280,11 @@ typedef struct RTSTRCACHEINT
     /** The total size of all the big entries. */
     size_t                  cbBigEntries;
     /** Hash collisions. */
-    size_t                  cHashCollisions;
+    uint32_t                cHashCollisions;
     /** Secondary hash collisions. */
-    size_t                  cHashCollisions2;
+    uint32_t                cHashCollisions2;
     /** The number of inserts to compare cHashCollisions to. */
-    size_t                  cHashInserts;
+    uint32_t                cHashInserts;
     /** The number of rehashes. */
     uint32_t                cRehashes;
     /** @} */
